+++ projects/ng-gallery/src/lib/components/gallery-core.component.ts
@@ -36,6 +36,10 @@ import { GalleryConfig } from '../models/config.model';
       <gallery-counter *ngIf="config.counter"
                        [state]="state">
       </gallery-counter>
+
+      <div class="g-box-template" style="position: absolute">
+        <ng-container *ngTemplateOutlet="config.boxTemplate; context: { state: state, config: config }"></ng-container>
+      </div>
     </div>
   `
 })
@@ -78,4 +82,13 @@ export class GalleryCoreComponent {
     return this.config.counterPosition;
   }
 
+  /** Set gallery counter position */
+  @HostBinding('attr.slidingDisabled') get slidingDisabled(): boolean {
+    return this.config.slidingDisabled;
+  }
+
+  /** Set gallery counter position */
+  @HostBinding('attr.thumbSlidingDisabled') get thumbSlidingDisabled(): boolean {
+    return this.config.thumbSlidingDisabled;
+  }
 }