+++ projects/ng-gallery/src/lib/components/gallery-slider.component.ts
@@ -7,20 +7,19 @@ import {
   OnChanges,
   ViewChild,
   SimpleChanges,
-  Inject,
   NgZone,
   ElementRef,
   EventEmitter,
-  ChangeDetectionStrategy,
-  PLATFORM_ID
+  ChangeDetectionStrategy
 } from '@angular/core';
-import { isPlatformBrowser } from '@angular/common';
-import { BehaviorSubject, Observable, Subscription, fromEvent } from 'rxjs';
-import { map, tap, debounceTime } from 'rxjs/operators';
+import { Platform } from '@angular/cdk/platform';
+import { Subscription, fromEvent } from 'rxjs';
+import { tap, debounceTime } from 'rxjs/operators';
 import { GalleryState, GalleryError } from '../models/gallery.model';
 import { GalleryConfig } from '../models/config.model';
 import { SlidingDirection } from '../models/constants';
-import { SliderState, WorkerState } from '../models/slider.model';
+import { SliderAdapter, HorizontalAdapter, VerticalAdapter } from './adapters';
+import { SmoothScrollManager } from '../smooth-scroll';
 
 declare const Hammer: any;
 
@@ -28,15 +27,15 @@ declare const Hammer: any;
   selector: 'gallery-slider',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
-    <div #container class="g-items-container">
+    <div class="g-items-container">
       <div #slider class="g-slider">
-        <gallery-item *ngFor="let item of state.items; let i = index"
+        <gallery-item *ngFor="let item of state.items; trackBy: trackByFn; index as i"
                       [type]="item.type"
                       [config]="config"
                       [data]="item.data"
                       [currIndex]="state.currIndex"
                       [index]="i"
-                      (tapClick)="itemClick.emit(i)"
+                      (click)="itemClick.emit(i)"
                       (error)="error.emit({itemIndex: i, error: $event})">
         </gallery-item>
       </div>
@@ -46,20 +45,14 @@ declare const Hammer: any;
 })
 export class GallerySliderComponent implements OnInit, OnChanges, OnDestroy {
 
-  /** Sliding worker */
-  private readonly _slidingWorker$ = new BehaviorSubject<WorkerState>({ value: 0, instant: true });
-
   /** HammerJS instance */
   private _hammer: any;
 
-  /** Subscription reference to window resize stream */
-  private _resizeSub$: Subscription;
+  /** Subscription reference to slider scroll stream */
+  private _scrollSub$: Subscription;
 
-  /** Subscription reference to slider state stream */
-  private _sliderStateSub$: Subscription;
-
-  /** Stream that emits sliding state */
-  sliderState$: Observable<SliderState>;
+  /** Slider adapter */
+  adapter: SliderAdapter;
 
   /** Gallery state */
   @Input() state: GalleryState;
@@ -80,37 +73,90 @@ export class GallerySliderComponent implements OnInit, OnChanges, OnDestroy {
 
   @ViewChild('slider', { static: true }) sliderEl: ElementRef;
 
-  get container(): HTMLElement {
-    return this.containerEl.nativeElement;
-  }
-
   get slider(): HTMLElement {
     return this.sliderEl.nativeElement;
   }
 
-  /** Item zoom */
-  get zoom() {
-    return { transform: `perspective(50px) translate3d(0, 0, ${ -this.config.zoomOut }px)` };
+  constructor(private _el: ElementRef,
+              private _zone: NgZone,
+              private _platform: Platform,
+              private _smoothScroll: SmoothScrollManager) {
   }
 
-  constructor(private _el: ElementRef, private _zone: NgZone, @Inject(PLATFORM_ID) private _platform: Object) {
+  ngOnChanges(changes: SimpleChanges): void {
+    if (changes.config) {
+      if (changes.config.currentValue?.slidingDirection !== changes.config.previousValue?.slidingDirection) {
+        switch (this.config.slidingDirection) {
+          case SlidingDirection.Horizontal:
+            this.adapter = new HorizontalAdapter(this.slider, this.config);
+            break;
+          case SlidingDirection.Vertical:
+            this.adapter = new VerticalAdapter(this.slider, this.config);
+            break;
+        }
 
-    // Activate sliding worker
-    this.sliderState$ = this._slidingWorker$.pipe(map((state: WorkerState) => ({
-      style: this.getSliderStyles(state),
-      instant: state.instant
-    })));
-  }
+        if (!changes.config.firstChange) {
+          // Keep the correct sliding position when direction changes
+          requestAnimationFrame(() => {
+            this.scrollToIndex(this.state.currIndex, 'auto');
+          });
+        }
 
-  ngOnChanges(changes: SimpleChanges) {
-    // Refresh the slider
-    if (changes.state) {
-      this.updateSlider({ value: 0, instant: changes.state.firstChange });
+        // Reactivate gestures
+        this.enableDisableGestures();
+      }
+
+      if (!changes.config.firstChange && changes.config.currentValue?.mouseSlidingDisabled !== changes.config.previousValue?.mouseSlidingDisabled) {
+        this.enableDisableGestures();
+      }
     }
 
-    // Enable/Disable gestures on changes
-    if (changes.config && changes.config.currentValue?.gestures !== changes.config.previousValue?.gestures) {
-      if (this.config.gestures) {
+    // Scroll to current index
+    if (changes.state && changes.state.currentValue?.currIndex !== changes.state.previousValue?.currIndex) {
+      requestAnimationFrame(() => {
+        this.scrollToIndex(this.state.currIndex, changes.state.firstChange ? 'auto' : 'smooth');
+      });
+    }
+  }
+
+  ngOnInit(): void {
+    this._zone.runOutsideAngular(() => {
+      // Subscribe to slider scroll event
+      this._scrollSub$ = fromEvent(this.slider, 'scroll', { passive: true }).pipe(
+        debounceTime(50),
+        tap(() => {
+          const index: number = this.adapter.measureIndex;
+          // Check if index value is
+          if (Number.isSafeInteger(index)) {
+            this._zone.run(() => this.action.emit(index));
+          }
+        })
+      ).subscribe();
+    });
+  }
+
+  ngOnDestroy(): void {
+    this.deactivateGestures();
+    this._scrollSub$?.unsubscribe();
+  }
+
+  trackByFn(index: number, item: any) {
+    return item.type;
+  }
+
+  private scrollToIndex(value: number, behavior: ScrollBehavior): void {
+    this._zone.runOutsideAngular(() => {
+      this.slider.style.scrollSnapType = 'unset';
+      this._smoothScroll.scrollTo(this.slider, this.adapter.getScrollToValue(value, behavior)).then(() => {
+        this.slider.style.scrollSnapType = this.adapter.scrollSnapType;
+      });
+    });
+  }
+
+  private enableDisableGestures(): void {
+    // Enable/Disable mouse sliding on desktop browser only
+    if (!this._platform.IOS && !this._platform.ANDROID) {
+      if (!this.config.mouseSlidingDisabled) {
         this.activateGestures();
       } else {
         this.deactivateGestures();
@@ -118,153 +164,57 @@ export class GallerySliderComponent implements OnInit, OnChanges, OnDestroy {
     }
   }
 
-  ngOnInit() {
-    this._zone.runOutsideAngular(() => {
-      // Set styles manually avoid triggering change detection on dragging
-      this._sliderStateSub$ = this.sliderState$.pipe(
-        tap((state: SliderState) => {
-          this.slider.style.transform = state.style.transform;
-          this.slider.style.height = state.style.height;
-          this.slider.style.width = state.style.width;
-          this.slider.classList.toggle('g-no-transition', state.instant);
-          this.container.style.transform = this.zoom.transform;
-        })
-      ).subscribe();
-    });
-
-    // Rearrange slider on window resize
-    if (isPlatformBrowser(this._platform)) {
-      this._resizeSub$ = fromEvent(window, 'resize').pipe(
-        debounceTime(200),
-        tap(() => this.updateSlider({ value: 0, instant: true }))
-      ).subscribe();
-    }
-  }
-
-  ngOnDestroy() {
-    this.deactivateGestures();
-    this._resizeSub$?.unsubscribe();
-    this._sliderStateSub$?.unsubscribe();
-    this._slidingWorker$.complete();
-  }
-
-  private activateGestures() {
+  private activateGestures(): void {
     if (typeof Hammer !== 'undefined') {
-      let direction: number;
-      let touchAction: 'pan-x' | 'pan-y' | 'compute' = 'compute';
+      // Destroy hammer if a previous instance was created
+      this.deactivateGestures();
 
-      if (this.config.slidingDirection === SlidingDirection.Horizontal) {
-        direction = Hammer.DIRECTION_HORIZONTAL;
-        if (this.config.reserveGesturesAction) {
-          touchAction = 'pan-x';
-        }
-      } else {
-        direction = Hammer.DIRECTION_VERTICAL;
-        if (this.config.reserveGesturesAction) {
-          touchAction = 'pan-y';
-        }
-      }
+      const direction: number = this.adapter.panDirection;
 
       // Activate gestures
-      this._hammer = new Hammer(this._el.nativeElement, { touchAction });
-      this._hammer.get('pan').set({ direction });
-
       this._zone.runOutsideAngular(() => {
-        this._hammer.on('pan', (e) => {
-          switch (this.config.slidingDirection) {
-            case SlidingDirection.Horizontal:
-              if (e.isFinal) {
-                this.updateSlider({ value: 0, instant: false });
-                this.horizontalPan(e);
-              } else {
-                this.updateSlider({ value: e.deltaX, instant: true });
-              }
-              break;
-            case SlidingDirection.Vertical:
-              if (e.isFinal) {
-                this.updateSlider({ value: 0, instant: false });
-                this.verticalPan(e);
-              } else {
-                this.updateSlider({ value: e.deltaY, instant: true });
-              }
-          }
+        this._hammer = new Hammer(this._el.nativeElement, { inputClass: Hammer.MouseInput });
+        this._hammer.get('pan').set({ direction });
+
+        let panOffset: number;
+
+        // Set panOffset for sliding on pan start event
+        this._hammer.on('panstart', () => {
+          this._smoothScroll.dismissOngoingScroll(this.slider);
+          this.slider.classList.add('g-sliding');
+          panOffset = this.adapter.scrollValue;
+          // Disable scroll-snap-type functionality
+          this.slider.style.scrollSnapType = 'unset';
+        });
+
+        this._hammer.on('panmove', (e: any) => this.slider.scrollTo(this.adapter.getPanValue(panOffset, e, 'auto')));
+
+        this._hammer.on('panend', (e: any) => {
+          this.slider.classList.remove('g-sliding');
+          this.onPanEnd(e);
         });
       });
     }
   }
 
-  private deactivateGestures() {
+  private deactivateGestures(): void {
     this._hammer?.destroy();
   }
 
-  /**
-   * Convert sliding state to styles
-   */
-  private getSliderStyles(state: WorkerState): any {
-    switch (this.config.slidingDirection) {
-      case SlidingDirection.Horizontal:
-        return {
-          transform: `translate3d(${ -(this.state.currIndex * this._el.nativeElement.offsetWidth) + state.value }px, 0, 0)`,
-          width: `calc(100% * ${ this.state.items.length })`,
-          height: '100%'
-        };
-      case SlidingDirection.Vertical:
-        return {
-          transform: `translate3d(0, ${ -(this.state.currIndex * this._el.nativeElement.offsetHeight) + state.value }px, 0)`,
-          width: '100%',
-          height: `calc(100% * ${ this.state.items.length })`,
-        };
-    }
-  }
-
-  private verticalPan(e) {
-    if (!(e.direction & Hammer.DIRECTION_UP && e.offsetDirection & Hammer.DIRECTION_VERTICAL)) {
-      return;
-    }
-    if (e.velocityY > 0.3) {
-      this.prev();
-    } else if (e.velocityY < -0.3) {
-      this.next();
-    } else {
-      if (e.deltaY / 2 <= -this._el.nativeElement.offsetHeight * this.state.items.length / this.config.panSensitivity) {
-        this.next();
-      } else if (e.deltaY / 2 >= this._el.nativeElement.offsetHeight * this.state.items.length / this.config.panSensitivity) {
-        this.prev();
-      } else {
-        this.action.emit(this.state.currIndex);
+  protected onPanEnd(e): void {
+    this._zone.run(() => {
+      const delta: number = this.adapter.getPanDelta(e);
+      const velocity: number = this.adapter.getPanVelocity(e);
+      // Check if scrolled item is great enough to navigate
+      if (Math.abs(delta) > this.adapter.clientSize / 2) {
+        return this.action.emit(delta > 0 ? 'prev' : 'next');
       }
-    }
-  }
-
-  private horizontalPan(e) {
-    if (!(e.direction & Hammer.DIRECTION_HORIZONTAL && e.offsetDirection & Hammer.DIRECTION_HORIZONTAL)) {
-      return;
-    }
-    if (e.velocityX > 0.3) {
-      this.prev();
-    } else if (e.velocityX < -0.3) {
-      this.next();
-    } else {
-      if (e.deltaX / 2 <= -this._el.nativeElement.offsetWidth * this.state.items.length / this.config.panSensitivity) {
-        this.next();
-      } else if (e.deltaX / 2 >= this._el.nativeElement.offsetWidth * this.state.items.length / this.config.panSensitivity) {
-        this.prev();
-      } else {
-        this.action.emit(this.state.currIndex);
+      // Check if velocity is great enough to navigate
+      if (Math.abs(velocity) > 0.3) {
+        return this.action.emit(velocity > 0 ? 'prev' : 'next');
       }
-    }
-  }
-
-  private next() {
-    this.action.emit('next');
-  }
-
-  private prev() {
-    this.action.emit('prev');
-  }
-
-  private updateSlider(state: WorkerState) {
-    const newState: WorkerState = { ...this._slidingWorker$.value, ...state };
-    this._slidingWorker$.next(newState);
+      this.scrollToIndex(this.state.currIndex, 'smooth');
+      this.action.emit(this.state.currIndex);
+    });
   }
 }