+++ projects/ng-gallery/src/lib/utils/gallery.default.ts
@@ -5,9 +5,9 @@ import {
   LoadingStrategy,
   SlidingDirection,
   ThumbnailsPosition,
-  ImageLoaderMode,
   DotsPosition,
-  CounterPosition
+  CounterPosition,
+  ThumbnailsView
 } from '../models/constants';
 import { GalleryState } from '../models/gallery.model';
 import { GalleryConfig } from '../models/config.model';
@@ -25,25 +25,33 @@ export const defaultState: GalleryState = {
 export const defaultConfig: GalleryConfig = {
   nav: true,
   loop: true,
-  zoomOut: 0,
   dots: false,
   thumb: true,
   dotsSize: 30,
   counter: true,
-  gestures: true,
   autoPlay: false,
   thumbWidth: 120,
   thumbHeight: 90,
-  panSensitivity: 25,
   disableThumb: false,
+  slidingDisabled: false,
+  thumbSlidingDisabled: false,
+  mouseSlidingDisabled: false,
+  thumbMouseSlidingDisabled: false,
   playerInterval: 3000,
-  reserveGesturesAction: false,
+  slidingDuration: 468,
+  slidingEase: {
+    x1: 0.42,
+    y1: 0,
+    x2: 0.58,
+    y2: 1
+  },
+  resizeDebounceTime: 50,
   imageSize: ImageSize.Contain,
-  thumbMode: ThumbnailsMode.Strict,
+  thumbView: ThumbnailsView.Default,
   dotsPosition: DotsPosition.Bottom,
   counterPosition: CounterPosition.Top,
   thumbPosition: ThumbnailsPosition.Bottom,
-  loadingStrategy: LoadingStrategy.Default,
+  loadingStrategy: LoadingStrategy.Preload,
   slidingDirection: SlidingDirection.Horizontal,
   navIcon: `<?xml version="1.0" encoding="UTF-8"?><svg width="512px" height="512px" enable-background="new 0 0 240.823 240.823" version="1.1" viewBox="0 0 240.823 240.823" xml:space="preserve" xmlns="http://www.w3.org/2000/svg"><path d="m183.19 111.82l-108.3-108.26c-4.752-4.74-12.451-4.74-17.215 0-4.752 4.74-4.752 12.439 0 17.179l99.707 99.671-99.695 99.671c-4.752 4.74-4.752 12.439 0 17.191 4.752 4.74 12.463 4.74 17.215 0l108.3-108.26c4.68-4.691 4.68-12.511-0.012-17.19z" fill="#fff"/></svg>`,
   loadingIcon: `<?xml version="1.0" encoding="UTF-8"?><svg stroke="#fff" viewBox="0 0 44 44" xmlns="http://www.w3.org/2000/svg"><g fill="none" fill-rule="evenodd" stroke-width="2"><circle cx="22" cy="22" r="1"><animate attributeName="r" begin="0s" calcMode="spline" dur="1.8s" keySplines="0.165, 0.84, 0.44, 1" keyTimes="0; 1" repeatCount="indefinite" values="1; 20"/><animate attributeName="stroke-opacity" begin="0s" calcMode="spline" dur="1.8s" keySplines="0.3, 0.61, 0.355, 1" keyTimes="0; 1" repeatCount="indefinite" values="1; 0"/></circle><circle cx="22" cy="22" r="1"><animate attributeName="r" begin="-0.9s" calcMode="spline" dur="1.8s" keySplines="0.165, 0.84, 0.44, 1" keyTimes="0; 1" repeatCount="indefinite" values="1; 20"/><animate attributeName="stroke-opacity" begin="-0.9s" calcMode="spline" dur="1.8s" keySplines="0.3, 0.61, 0.355, 1" keyTimes="0; 1" repeatCount="indefinite" values="1; 0"/></circle></g></svg>`